// src/util/graccResponse.d.ts
// Auto-generated by https://transform.tools/json-to-typescript
// by taking response body of a query to Elasticsearch

export interface Root {
  took: number;
  timed_out: boolean;
  _shards: ShardInfo;
  hits: HitInfo;
  aggregations: Aggregations;
}

export interface ShardInfo {
  total: number;
  successful: number;
  skipped: number;
  failed: number;
}

export interface HitInfo {
  total: Total;
  max_score: number;
  hits: Hit[];
}

export interface Total {
  value: number;
  relation: string;
}

export interface Hit {
  _index: string;
  _id: string;
  _score: number;
  _source: Job;
}

export interface Job {
  OIM_WLCGAPELNormalFactor?: number;
  CoreHours: number;
  WallDuration: number;
  OIM_ResourceGroup?: string;
  type: string;
  OIM_Resource?: string;
  OIM_Match?: string;
  DN: string;
  VOName: string;
  OIM_WLCGAPELNormalizedWallDuration?: number;
  CommonName: string;
  event: JobEvent;
  "@received": string;
  RawCorrectedReportableVOName: string;
  CpuDuration: number;
  OIM_Site: string;
  OIM_NSFFieldOfScience: string;
  Processors: number;
  OIM_WLCGAccountingName?: string;
  Count: number;
  Grid: string;
  EndTime: string;
  OIM_UsageModel?: string;
  GPUs: string;
  CpuDuration_system: number;
  checksum: string;
  CpuDuration_user: number;
  ReportableVOName: string;
  "@version": string;
  GPUHours: number;
  RawProjectName: string;
  "@timestamp": string;
  ResourceType: string;
  ProbeName: string;
  OIM_WLCGAPELNormalizedCpuDuration?: number;
  OIM_Facility?: string;
  OIM_FieldOfScience: string;
  OIM_ID?: string;
  Host_description: string;
  Resource_ExitCode: number;
  ProjectName: string;
  SiteName: string;
  Njobs: number;
  RawVOName: string;
  OIM_FQDN?: string;
  CorrectedReportableVOName?: string;
  RawOIM_Site: string;
  OIM_PIName?: string;
  OIM_Organization?: string;
  OIM_Department?: string;
}

export interface JobEvent {
  original: string;
}

export interface Aggregations {
  EndTime: EndTime;
}

export interface EndTime {
  buckets: Bucket[];
}

export interface Bucket {
  key_as_string: string;
  key: number;
  doc_count: number;
  CoreHours: CoreHours;
  Njobs: Njobs;
}

export interface CoreHours {
  value: number;
}

export interface Njobs {
  value: number;
}